refactor: rename use-mobile.ts to useMobile.ts for consistency
